Safari-609.1.11.1/LayoutTests/fast/speechsynthesis/speech-synthesis-elapsed-time-expected.txt

This tests that elapsed time is a number that is reasonable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Our start time for speech should be very small.
PASS startTime < .5 is true
Word boundary received: startTime should be less than the current elapsed time.
PASS startTime < event.elapsedTime is true
Word boundary received: startTime should be less than the current elapsed time.
PASS startTime < event.elapsedTime is true
The finished elapsed time should be small and the startTime should have occurred before the end.
PASS startTime < event.elapsedTime is true
PASS event.elapsedTime < 1 is true
PASS successfullyParsed is true

TEST COMPLETE
